Numeric controllers operate on int only

+++ shortcut_composer/templates/pie_menu_utils/pie_settings_impl/dispatch_pie_settings.py
@@ -17,7 +17,7 @@ def dispatch_pie_settings(controller: Controller) -> Type[PieSettings]:
     """Return the right settings type based on value type."""
     if issubclass(controller.TYPE, str):
         return PresetPieSettings
-    elif issubclass(controller.TYPE, float):
+    elif issubclass(controller.TYPE, int):
         return NumericPieSettings
     elif issubclass(controller.TYPE, EnumGroup):
         return EnumGroupPieSettings
